+++ packages/xchain-doge/__mocks__/sochain.ts
@@ -1,7 +1,4 @@
-import axios from 'axios'
-import MockAdapter from 'axios-mock-adapter'
-
-const mock = new MockAdapter(axios)
+import mock from './axios-adapter'
 
 type MockConfig = {
   url?: string
@@ -48,14 +45,15 @@ export default {
       return [200, resp]
     })
 
-    //Mock ltc node send tx
-    mock.onPost(/https:\/\/testnet.ltc.thorchain.info/).reply(function () {
+    // Mock send tx
+    mock.onPost(/\/send_tx/).reply(function () {
       return [
         200,
         {
-          id: '1',
-          result: 'TEST_OK',
-          error: null,
+          data: {
+            network: 'DOGE',
+            txid: 'mock-txid-sochain',
+          },
         },
       ]
     })