+++ packages/xchain-doge/src/index.ts
@@ -3,6 +3,7 @@ export * from './client'
 export * from './const'
 export {
   broadcastTx,
+  buildTx,
   getDefaultFees,
   getDefaultFeesWithRates,
   validateAddress,
@@ -10,4 +11,5 @@ export {
   scanUTXOs,
   getPrefix,
 } from './utils'
+export { getSendTxUrl } from './blockcypher-api'
 export { createTxInfo } from './ledger'